Add SGLang true-on-policy namespace
