+++ python/sglang/srt/model_executor/model_runner.py
@@ -116,7 +116,7 @@ from sglang.srt.layers.moe.routed_experts_capturer import (
     get_global_experts_capturer,
     set_global_experts_capturer,
 )
-from sglang.srt.layers.on_policy_utils import is_tp_invariant_target
+from sglang.srt.true_on_policy import is_tp_invariant_target
 from sglang.srt.layers.pooler import EmbeddingPoolerOutput
 from sglang.srt.layers.quantization.fp8_kernel import fp8_dtype
 from sglang.srt.layers.sampler import create_sampler
