+++ python/sglang/srt/models/qwen3.py
@@ -15,7 +15,7 @@ from sglang.srt.layers.dp_attention import get_attention_tp_rank, get_attention_
 from sglang.srt.layers.layernorm import RMSNorm
 from sglang.srt.layers.linear import QKVParallelLinear, RowParallelLinear
 from sglang.srt.layers.logits_processor import LogitsProcessor
-from sglang.srt.layers.on_policy_utils import (
+from sglang.srt.true_on_policy import (
     get_on_policy_rms_norm_kwargs,
     should_force_bfloat16_dense_tensor_math,
 )
